+++ webclient/app/src/app/features/TrajectoryMap.jsx
@@ -3,10 +3,10 @@ import { useTranslation } from 'react-i18next';
 import StreamRest from "../services/StreamRest";
 import WebSocketClient from "../services/WebSocketClient";
 import LiveMapView from "./LiveMapView";
-import { Box, Card, IconButton, Stack } from "@mui/material";
+import { Box, Card, IconButton, Stack, Typography } from "@mui/material";
 import PlayCircleFilledWhiteIcon from '@mui/icons-material/PlayCircleFilledWhite';
 import StopCircleIcon from '@mui/icons-material/StopCircle';
-
+import ColorFunctions from "../services/ColorFunctions";
 
 function TrajectoryMap() {
     const { t } = useTranslation();
@@ -16,11 +16,13 @@ function TrajectoryMap() {
     const [streams, setStreams] = useState({});
     const [markerList, setMarkerList] = useState({});
     const [showMap, setShowMap] = useState(false);
+    const [started, setStarted] = useState(false);
+    const colorFunctions = useRef(new ColorFunctions());
 
     useEffect(() => {
         streamRest.getAvailableStreams().then(response => {
             const streams = response.data;
-            const colors = generateDistinctColors(Object.keys(response.data).length);
+            const colors = colorFunctions.current.generateDistinctColors(Object.keys(response.data).length);
             let streamsAndColors = {}
             streams.forEach((stream, index) => {
                 console.log(stream);
@@ -28,56 +30,12 @@ function TrajectoryMap() {
             });
             setStreams(streamsAndColors);
             setShowMap(true);
-    
+
             wsClient.current.setup(handleMessage, streams);
         });
         return () => wsClient.current.disconnect();
     }, []);
 
-    function generateDistinctColors(n) {
-        const colors = [];
-        // Use golden ratio to help spread the hues evenly
-        const goldenRatio = 0.618033988749895;
-        let hue = Math.random(); // Start at random hue
-    
-        for (let i = 0; i < n; i++) {
-            hue = (hue + goldenRatio) % 1; // Use golden ratio to increment hue
-            
-            // Convert HSV to RGB
-            // Using fixed Saturation (100%) and Value (100%) for vibrant colors
-            const rgb = hsvToRgb(hue, 1, 1);
-            
-            // Scale to your desired range (0-100 in this case)
-            colors.push([
-                Math.round(rgb[0] * 100),
-                Math.round(rgb[1] * 100),
-                Math.round(rgb[2] * 100)
-            ]);
-        }
-        
-        return colors;
-    }
-    
-    function hsvToRgb(h, s, v) {
-        let r, g, b;
-        const i = Math.floor(h * 6);
-        const f = h * 6 - i;
-        const p = v * (1 - s);
-        const q = v * (1 - f * s);
-        const t = v * (1 - (1 - f) * s);
-    
-        switch (i % 6) {
-            case 0: r = v; g = t; b = p; break;
-            case 1: r = q; g = v; b = p; break;
-            case 2: r = p; g = v; b = t; break;
-            case 3: r = p; g = q; b = v; break;
-            case 4: r = t; g = p; b = v; break;
-            case 5: r = v; g = p; b = q; break;
-        }
-    
-        return [r, g, b];
-    }     
-    
     function handleMessage(trackedObjectList, streamId) {
         let newMarkers = [];
         trackedObjectList.forEach(trackedObject => {
@@ -93,50 +51,44 @@ function TrajectoryMap() {
             } else {
             }
         });
-        setMarkerList(prevMarkerList => ({...prevMarkerList, [streamId]: newMarkers}));
+        setMarkerList(prevMarkerList => ({ ...prevMarkerList, [streamId]: newMarkers }));
     }
 
     function startStream() {
         wsClient.current.connect();
+        setStarted(true);
     }
 
     function stopStream() {
         wsClient.current.disconnect();
+        setStarted(false);
     }
 
     return (
         <>
-            <Stack direction="column" spacing={2}>
-                <div>
-                    <h1>{t('trajectory.title')}</h1>
-                </div>
-                <Card>
-                    <Stack direction="row" spacing={2}>
-                        <div>
-                            <IconButton onClick={startStream}><PlayCircleFilledWhiteIcon /></IconButton>
-                        </div>
-                        <div>
-                            <IconButton onClick={stopStream}><StopCircleIcon /></IconButton>
-                        </div>
-                    </Stack>
-                </Card>
-                <Box sx={{
-                    border: 1,
-                    height: 850,
-                    width: "auto",
-                    m: 5,
-                    position: 'relative'
-                }}>
-                    {showMap ? (
-                        <LiveMapView
-                            streams={streams}
-                            markerList={markerList}
-                        />
-                    ) : (
-                        <></>
-                    )}
-                </Box>
-            </Stack>
+            {showMap ? (
+                <LiveMapView
+                    streams={streams}
+                    markerList={markerList}
+                />
+            ) : (
+                <></>
+            )}
+            <Box sx={{
+                position: 'fixed',
+                top: 60,
+                left: 10,
+                width: '100%'
+            }}>
+                <Stack direction="row" spacing={5} width="fullWidth">
+                    {!started ?
+                        <IconButton size="large" color="primary" onClick={startStream}><PlayCircleFilledWhiteIcon /></IconButton> :
+                        <IconButton size="large" color="error" onClick={stopStream}><StopCircleIcon /></IconButton>}
+                    <Typography variant="h1">
+                        {t('map.title')}
+                    </Typography>                        
+                </Stack>
+            </Box>
         </>
     )
 }
